fix: 홈페이지 비지니스 로직 분리
