+++ app/(GNB-layout)/page.tsx
@@ -6,33 +6,13 @@ import SubscriptionSection from "../main/subscription-section";
 import HistorySection from "../main/history-section";
 import SimulationSection from "../main/simulation.section";
 import LazyLoad from "@/hooks/main/lazy-loading";
-import { useEffect, useState } from "react";
+
 import Loading from "@/components/shared/loading/loading";
 import BackgroundSection from "../main/background-section";
-import {
-  AMPLITUDE_EVENT_DISPLAY_NAME,
-  AMPLITUDE_EVENT_LOG_NAME,
-} from "@/constants/amplitude";
-import useTrackPageView from "@/hooks/amplitude/ust-track-page-view-with-UTM";
+import useMain from "@/hooks/main/use-main";
 
 const Home = () => {
-  const { trackPageView } = useTrackPageView();
-  const [loading, setLoading] = useState(true);
-
-  useEffect(() => {
-    if (typeof window === "undefined") return;
-    trackPageView(
-      AMPLITUDE_EVENT_LOG_NAME.LANDING_PAGE_VIEW,
-      AMPLITUDE_EVENT_DISPLAY_NAME.LANDING_PAGE_VIEW,
-    );
-  }, [trackPageView]);
-
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      setLoading(false);
-    }, 350);
-    return () => clearTimeout(timer);
-  }, []);
+  const { loading } = useMain();
 
   return (
     <>
